Create split.py
Write a Python program that takes an input from the user, then creates a new list excluding the first and last two elements of the given list and prints the new list. If there are not enough elements in the list to do the task, print "Not possible".
Note: You may use list slicing.