Create modify_list.py
Write a Python program that reads a string as an input from the user where multiple numbers are separated by spaces. Then, make a list of numbers from the input string without using the split() function and print the list. Finally, remove all the occurrences of even numbers from the same input list and print the modified list.