Create reverse_list.py
Take a list of 5 elements as user input and print the list in reverse order. Slicing can be used handling the given list as input. But to reverse the list you must not use slicing and split( ). Hint: You may use a loop to show the reversed list as a new list.