Create list_in_list.py
Assume you have a class of five students. And as a faculty you need to enter their name and CSE110 GPA in a list called grade_list. This particular grade_list consists of individual lists of the student's name and GPA. Hence, it makes the grade_list a nested list / list of lists.
Particular student information will be given as input by the faculty. Please refer to the example below to design your program.